Selection of Levels
Gave user the choice of level/stage. Changed speed and update to double so that we can speed up the pace of the stage without it being drastic

+++ csc121-creative-project/src/Platform.java
@@ -41,7 +41,7 @@ public class Platform {
     }
 	
 	// Updates the platform's position
-	public Platform update(int speed) {
+	public Platform update(double speed) {
 		this.p.x += speed;
 		return new Platform(this.p, this.width, this.height, this.color);
 	}
